changed using CDN url from environment

+++ app/actions/species.js
@@ -11,7 +11,9 @@ import {
 } from '../utils/api';
 import { LogTypes } from '../utils/constants';
 import { SET_SPECIE, CLEAR_SPECIE } from './Types';
-import { getCdnUrls } from './user';
+import { APIConfig } from './Config';
+
+const { protocol, cdnUrl } = APIConfig;
 
 /**
  * This function dispatches type SET_SPECIE with payload specie to show specie detail on SpecieInfo screen
@@ -220,30 +222,28 @@ export const UpdateSpeciesImage = (image, speciesId, SpecieGuid) => {
 
 export const getBase64ImageFromURL = async (specieImage) => {
   return new Promise((resolve) => {
-    getCdnUrls(i18next.language)
-      .then((cdnMedia) => {
-        RNFS.downloadFile({
-          fromUrl: `${cdnMedia.cache}/species/default/${specieImage}`,
-          toFile: `${RNFS.DocumentDirectoryPath}/${specieImage}`,
-        }).promise.then((response) => {
-          if (response.statusCode === 200) {
-            RNFS.readFile(`${RNFS.DocumentDirectoryPath}/${specieImage}`, 'base64')
-              .then((data) => {
-                resolve(data);
-                RNFS.unlink(`${RNFS.DocumentDirectoryPath}/${specieImage}`).catch((err) => {
-                  // `unlink` will throw an error, if the item to unlink does not exist
-                  console.error(err.message);
-                });
-              })
-              .catch((err) => console.error('Error while reading file image'));
-          } else {
-            resolve();
-          }
-        });
-      })
-      .catch((err) => {
-        bugsnag.notify(err);
-        resolve();
+    if (cdnUrl) {
+      RNFS.downloadFile({
+        fromUrl: `${protocol}://${cdnUrl}/media/cache/species/default/${specieImage}`,
+        toFile: `${RNFS.DocumentDirectoryPath}/${specieImage}`,
+      }).promise.then((response) => {
+        if (response.statusCode === 200) {
+          RNFS.readFile(`${RNFS.DocumentDirectoryPath}/${specieImage}`, 'base64')
+            .then((data) => {
+              resolve(data);
+              RNFS.unlink(`${RNFS.DocumentDirectoryPath}/${specieImage}`).catch((err) => {
+                // `unlink` will throw an error, if the item to unlink does not exist
+                console.error(err.message);
+              });
+            })
+            .catch((err) => console.error('Error while reading file image'));
+        } else {
+          resolve();
+        }
       });
+    } else {
+      bugsnag.notify(err);
+      resolve();
+    }
   });
 };
